rename(route-folder): 코멘트 반영
