+++ src/app/(admin)/admin/notice/write/[id]/_components/NoticeEditForm/NoticeEditForm.tsx
@@ -9,7 +9,7 @@ import {
 import { WriteActionSection } from "@/components";
 import { NoticeEditFormValues } from "../../_types/NoticeEditFormValues";
 import NoticeEditImageSection from "./_internal/NoticeEditImageSection";
-import { useSubmitNoticeEdit } from "../../../_hooks/useSubmitNoticeEdit";
+import { useSubmitNoticeEdit } from "../../_hooks/useSubmitNoticeEdit";
 
 const IMAGE_HELP_TEXT = "*사진은 최대 5장 첨부가 가능합니다. (선택)";
